LayoutTests/accessibility/anchor-link-selection-and-focus.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../fast/js/resources/js-test-style.css">
<script>
var successfullyParsed = false;
</script>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">
<p><a id="link" href="#anchor">A link pointing to an anchor at the end of the document</a></p>
<p>A paragraph with a some text in the middle</p>
<p><a id="anchor" name="anchor" href="">The Anchor</a></p>
<p id="description"></p>
<div id="console"></div>
<script>
description("This tests that the caret position is properly updated when following anchor links.");

if (window.layoutTestController) {
    layoutTestController.dumpAsText();

    if (window.accessibilityController) {
        var link = document.getElementById("link");
        var anchor = document.getElementById("anchor");

        // Get needed objects.
        link.focus();
        var axLink = accessibilityController.focusedElement;
        debug("axLink.role is: '" + axLink.role + "'");
        anchor.focus();
        var axAnchor = accessibilityController.focusedElement;
        debug("axAnchor.role is: '" + axAnchor.role + "'");

        // Focus on the link and check initial state.
        link.focus();
        shouldBe("axLink.isFocused", "true");
        shouldBe("axAnchor.isFocused", "false");

        // Execute action and check state.
        axLink.press()

        // Check final state.
        shouldBe("getSelection().anchorNode.nodeValue", "'The Anchor'");
        shouldBe("axLink.isFocused", "false");
        shouldBe("axAnchor.isFocused", "true");
    }
}

successfullyParsed = true;
</script>
<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
